Refactor
Refactored store slices: actions, slice declaration and async thunks
moved to their own dirrectories
