+++ src/store/offers/slice.ts
@@ -1,27 +1,9 @@
-import {
-  type PayloadAction,
-  createSlice,
-  createAsyncThunk,
-} from '@reduxjs/toolkit';
+import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 
 import { Offer } from '@/types/offer';
 
-import { AppDispatch, RootState } from '@/types/store';
-import { AppApi } from '@/api';
-import { NameSpace } from '@/consts/namespace';
-import { toggleFavorite } from './favorites/async-actions';
-
-type PickedSlice = Pick<RootState, NameSpace.Offers>;
-
-export const fetchOffers = createAsyncThunk<
-  Offer[],
-  undefined,
-  {
-    dispatch: AppDispatch;
-    state: RootState;
-    extra: AppApi;
-  }
->('offers/fetchOffers', (_arg, { extra: api }) => api.offers.getOffers());
+import { toggleFavorite } from '../favorites/async-actions';
+import { fetchOffers } from './async-actions';
 
 type OffersState = {
   entities: Offer[];
@@ -69,18 +51,5 @@ export const offersSlice = createSlice({
 });
 
 export const { loaded: offersLoaded, activeOfferChanged } = offersSlice.actions;
-
-const filterOffersByCity = (offersList: Offer[], cityName: string) =>
-  offersList.filter((offer) => offer.city.name === cityName);
-
-export const offersSelector = {
-  entities: (state: PickedSlice) => state.offers.entities,
-  filterdByCity: (state: PickedSlice, name: string) => {
-    const offers = state.offers.entities;
-    return filterOffersByCity(offers, name);
-  },
-  isLoading: (state: PickedSlice) => state.offers.isLoading,
-  activeOffer: (state: PickedSlice) => state.offers.activeOffer,
-};
 
 export default offersSlice.reducer;
