+++ src/store/user/selectors.ts
@@ -1,6 +1,9 @@
+import { NameSpace } from '@/consts/namespace';
 import { RootState } from '@/types/store';
 
+type UserSlice = Pick<RootState, NameSpace.User>;
+
 export const userSelector = {
-  authorizationStatus: (state: RootState) => state.user.authorizationStatus,
-  info: (state: RootState) => state.user.info,
+  authorizationStatus: (state: UserSlice) => state.user.authorizationStatus,
+  info: (state: UserSlice) => state.user.info,
 };
